+++ car_parking/src/repository/users.py
@@ -98,61 +98,61 @@ async def change_password(user: User, new_password: str, db: Session) -> None:
     db.commit()
     db.refresh(user)
 
-
-async def update_avatar(email, url: str, db: Session) -> User:
-    """
-    The update_avatar function updates the avatar of a user.
+# not used
+# async def update_avatar(email, url: str, db: Session) -> User:
+#     """
+#     The update_avatar function updates the avatar of a user.
     
-    Args:
-        email (str): The email address of the user to update.
-        url (str): The URL for the new avatar image.
-        db (Session, optional): A database session object.
+#     Args:
+#         email (str): The email address of the user to update.
+#         url (str): The URL for the new avatar image.
+#         db (Session, optional): A database session object.
     
-    :param email: Find the user in the database
-    :param url: str: The URL for the new avatar image
-    :param db: Session: Pass the database session to the function
-    :return: A user object
-    """
-    user = await get_user_by_email(email, db)
-    user.avatar = url
-    db.commit()
-    db.refresh(user)
-    return user
+#     :param email: Find the user in the database
+#     :param url: str: The URL for the new avatar image
+#     :param db: Session: Pass the database session to the function
+#     :return: A user object
+#     """
+#     user = await get_user_by_email(email, db)
+#     user.avatar = url
+#     db.commit()
+#     db.refresh(user)
+#     return user
 
-
-async def get_user_by_id(user_id: int, db: Session) -> User | None:
-    """
-    The get_user_by_id function returns a User object from the database, given an id.
-        Args:
-            user_id (int): The id of the user to be retrieved.
-            db (Session): A Session instance for interacting with the database.
+# transfered to the repository/admin.py
+# async def get_user_by_id(user_id: int, db: Session) -> User | None:
+#     """
+#     The get_user_by_id function returns a User object from the database, given an id.
+#         Args:
+#             user_id (int): The id of the user to be retrieved.
+#             db (Session): A Session instance for interacting with the database.
     
-    :param user_id: int: Specify the type of parameter that is expected to be passed in
-    :param db: Session: Pass the database session to the function
-    :return: A user object or none
-    :doc-author: Trelent
-    """
-    return db.query(User).filter(User.id==user_id).first()
+#     :param user_id: int: Specify the type of parameter that is expected to be passed in
+#     :param db: Session: Pass the database session to the function
+#     :return: A user object or none
+#     :doc-author: Trelent
+#     """
+#     return db.query(User).filter(User.id==user_id).first()
 
-
-async def change_user_role(user: User, body: UserRoleUpdate, db: Session) -> User:
-    """
-    The change_user_role function changes the role of a user.
-        Args:
-            user (User): The User object to change the role of.
-            body (UserRoleUpdate): A UserRoleUpdate object containing the new role for this user.
-            db (Session): The database session to use when changing this users' role in our database.
+# transfered to the repository/admin.py
+# async def change_user_role(user: User, body: UserRoleUpdate, db: Session) -> User:
+#     """
+#     The change_user_role function changes the role of a user.
+#         Args:
+#             user (User): The User object to change the role of.
+#             body (UserRoleUpdate): A UserRoleUpdate object containing the new role for this user.
+#             db (Session): The database session to use when changing this users' role in our database.
     
-    :param user: User: Get the user object from the database
-    :param body: UserRoleUpdate: Get the role from the request body
-    :param db: Session: Access the database
-    :return: A user object with updated role
-    :doc-author: Trelent
-    """
-    user.role = body.role
-    db.commit()
-    db.refresh(user)
-    return user
+#     :param user: User: Get the user object from the database
+#     :param body: UserRoleUpdate: Get the role from the request body
+#     :param db: Session: Access the database
+#     :return: A user object with updated role
+#     :doc-author: Trelent
+#     """
+#     user.role = body.role
+#     db.commit()
+#     db.refresh(user)
+#     return user
     
     
 async def delete_user(user_id: int, db: Session) -> None:
@@ -169,31 +169,33 @@ async def delete_user(user_id: int, db: Session) -> None:
         db.commit()
     return None 
 
-async def get_imagis_quantity(user:User, db: Session):
-    """
-    The get_imagis_quantity function returns the number of images that a user has uploaded to the database.
-        Args:
-            user (User): The User object whose image quantity is being requested.
-            db (Session): The database session used for querying and updating data in the database.
+# async def get_imagis_quantity(user:User, db: Session):
+#     """
+#     The get_imagis_quantity function returns the number of images that a user has uploaded to the database.
+#         Args:
+#             user (User): The User object whose image quantity is being requested.
+#             db (Session): The database session used for querying and updating data in the database.
     
-    :param user:User: Get the user_id of the user
-    :param db: Session: Connect to the database and query it
-    :return: The quantity of images that a user has uploaded to the database
-    """
-    all_images = db.query(Image).filter_by(user_id = user.id).all()
-    quantity_of_loaded_images = len(all_images)
-    return quantity_of_loaded_images
+#     :param user:User: Get the user_id of the user
+#     :param db: Session: Connect to the database and query it
+#     :return: The quantity of images that a user has uploaded to the database
+#     """
+#     all_images = db.query(Image).filter_by(user_id = user.id).all()
+#     quantity_of_loaded_images = len(all_images)
+#     return quantity_of_loaded_images
 
-async def return_all_users(db: Session) -> dict:
-    """
-    The return_all_users function retrieves all usernames from the User table.
 
-    :param db: Session: Database session
-    :return: A dictionary containing all usernames from the User table
-    """
-    users = db.query(User).all()
-    usernames = {f"username(id: {user.id})": user.username for user in users}
-    return usernames
+# transfered to repsitory/admin
+# async def return_all_users(db: Session) -> dict:
+#     """
+#     The return_all_users function retrieves all usernames from the User table.
+
+#     :param db: Session: Database session
+#     :return: A dictionary containing all usernames from the User table
+#     """
+#     users = db.query(User).all()
+#     usernames = {f"username(id: {user.id})": user.username for user in users}
+#     return usernames
 
 async def update_banned_status(user: User, db: Session):
     """
