--- car_parking/src/repository_or/user_or.py
@@ -1,13 +1,0 @@
-#original
-
-# from typing import Type
-# from fastapi import HTTPException, status
-
-# from sqlalchemy.orm import Session, joinedload
-# from ..database.models import User
-
-
-# async def get_user_by_email(email: str, db: Session) -> User | None:
-#     return db.query(User).filter_by(email=email).first()
-
-
